Remove useless summaries from the release files
We used to put information about the current status of the series in
there, which lead to some summaries being repeated over and over. This
is no longer necessary, as we clearly explain what each series contains.
So if the release summary doesn't explain the content of this exact
release, it's useless.